# Booking/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Room Booking system
  version: 1.0.0
  description: >
    It is a  Use Case — Room Booking Management system for A co-working space manages room bookings through phone calls and a whiteboard.

  contact:
    name: Booking Manager 
    email: support@booking.email.com

servers:
  # - url: https://localhost:8000
  #   description: devlopment Server
  # - url: https://api.booking.com/v1
  #   description: Production Server
  # - url: https://api.booking.com/v1
  #   description: Staging Server
  - url: {BASE_URL}

  
paths:
  /api/v1/booking:
    post:
      tags:
        - Booking
      summary: Book the Meeting Room Avalaible 
      description: >
        
      operationId: chatWithBrewBuddy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              createBookingExample:
                summary: Customer requests a room booking
                value:
                  customer_id: 101
                  room_id: 5
                  booking_date: "2026-05-15"
                  start_time: "2026-05-15T09:00:00Z"
                  end_time: "2026-05-15T11:00:00Z"
      responses:
        '200':
          description: Successfully processed Booking request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                orderStatusOutput:
                  summary: Response for Booking request
                  value:
                    status: Success
                    code : 201
                    reply: "Room Booked Successfully"
                    error: null
                    request_id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                    timestamp: "2026-03-19T14:25:36.123456+00:00"
                  
        '400':
          description: Bad Request - The JSON is malformed or the request structure is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                message: "Invalid JSON payload. Check for missing braces or quotes."
                request_id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                timestamp: "2026-03-19T14:25:36.123456+00:00"

        '422':
          description: Unprocessable Content.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingFields:
                  summary: Missing required booking data
                  value:
                    status: 422
                    message: "Booking failed. Please provide room ID, date, and times."
                    request_id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                    timestamp: "2026-03-19T14:25:36.123456+00:00"
                emptyMessage:
                  summary: Invalid UUID  
                  value:
                    status: 422
                    message: "Please Enter the Correct Customer UUID "
                    request_id: "a1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p6"
                    timestamp: "2026-03-19T14:26:00.000000+00:00"

        '500':
          description: Internal server error while processing booking request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 500
                message: "Unexpected error occurred while processing booking request."
                request_id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                timestamp: "2026-03-19T14:25:36.123456+00:00"

  # ----------
  # /api/v1/rooms:
  #   get:
  #       tags:
  #         - Booking
  #       summary: Show Rooms Available for Booking
  #       description: Returns a list of rooms based on availability.
  #       parameters:
  #         - name: date
  #           in: query
  #           required: true
  #           description: Date to check room availability
  #           schema:
  #             type: string
  #             format: date
  #             example: "2026-05-11"

  #       responses:
  #         '200':
  #           description: Successful retrieval of available rooms
  #           content:
  #             application/json:
  #               schema:
  #                 allOf:
  #                   - $ref: '#/components/schemas/APIResponse'
  #                   - type: object
  #                     properties:
  #                       data:
  #                         type: array
  #                         items:
  #                           type: object
  #                           properties:
  #                             room_id:
  #                               type: integer
  #                             room_name:
  #                               type: string
  #                             status:
  #                               type: string
  #               example:
  #                 code: 200
  #                 status: "Success"
  #                 message: "Rooms retrieved successfully"
                 
  #                 error: null
  #                 data:
  #                   - room_id: 5
  #                     room_name: "CR-1"
  #                     status: "Available"
  #                   - room_id: 12
  #                     room_name: "MR-1"
  #                     status: "Available"
  #                 request_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
  #                 timestamp: "2026-05-15T17:34:00Z"

  /api/v1/bookings/{id}:
    get:
      tags:
        - Booking
      summary: Get all bookings For a Particular id
      description: Returns the details of a single booking based on its unique ID.
      operationId: getBookingById
      parameters:
        - name: id
          in: path
          description: The unique identifier of the booking to retrieve.
          required: true
          schema:
            type: string
            example: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
      responses:
        '200':
          description: OK - Booking details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 200
                status: "Success"
                message: "Booking records retrieved successfully"
                data:
                  - customer_id: 101
                    room_id: 5
                    booking_date: "2026-05-15"
                  - customer_id: 102
                    room_id: 3
                    booking_date: "2026-05-16"
                request_id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                timestamp: "2026-03-19T14:30:00Z"

        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 404
                status: "error"
                message: "No booking found with the provided ID."
                data: null
                request_id: "a1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p6"
                timestamp: "2026-03-19T14:30:00Z"

        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 422
                status: "error"
                message: "Validation failed"
                data: null
                error:
                  detail:
                    - loc: ["path", "id"]
                      msg: "invalid uuid format"
                      type: "value_error.invalid"
                request_id: "b1a2c3d4-e5f6-7890-abcd-1234567890ef"
                timestamp: "2026-03-20T10:15:30Z" 
#  put 
    put:
      tags:
        - Booking
      summary: Update an existing booking
      description: Updates the booking details for a specific ID. 
      operationId: updateBooking
      parameters:
        - name: id
          in: path
          description: The unique identifier of the booking to update.
          required: true
          schema:
            type: string
            format: uuid
            example: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              customer_id: 101
              room_id: 5
              booking_date: "2026-05-15"
              start_time: "2026-05-15T09:00:00Z"
              end_time: "2026-05-15T11:00:00Z"
      responses:
        '200':
          description: OK - Booking updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 200
                status: "Success"
                message: "Booking updated successfully"
                data:
                  booking_id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                  updated_at: "2026-03-20T11:00:00Z"
                request_id: "c9d8e7f6-a5b4-3210-fe09-876543210fed"
                timestamp: "2026-03-20T11:00:05Z"

        '400':
          description: Bad Request - Malformed JSON structure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 400
                status: "error"
                message: "Invalid JSON format."
                request_id: "d1e2f3a4-b5c6-7890-abcd-ef1234567890"
                timestamp: "2026-03-20T11:01:00Z"

        '404':
          description: Not Found - No booking exists with this ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 404
                status: "error"
                message: "Booking record not found."
                request_id: "e1f2a3b4-c5d6-7890-abcd-1234567890fe"
                timestamp: "2026-03-20T11:02:00Z"

        '422':
          description: Validation Error - Semantic errors in the request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 422
                status: "error"
                message: "Validation failed"
                error:
                  detail:
                    - loc: ["body", "room_id"]
                      msg: "Room is already booked for this time slot"
                      type: "value_error.conflict"
                request_id: "f1a2b3c4-d5e6-7890-abcd-0987654321fe"
                timestamp: "2026-03-20T11:03:00Z" 
    delete:
      tags:
        - Booking
      summary: Soft Delete a Particular Booking
      description: Marks a booking as deleted 
      operationId: softDeleteBooking
      parameters:
        - name: id
          in: path
          description: The unique identifier of the booking to soft delete.
          required: true
          schema:
            type: string
            format: uuid
            example: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
      # requestBody:
      #   description: Partial update to set the deletion flag or status.
      #   required: true
      #   content:
      #     application/json:
      #       schema:
      #         type: object
      #         properties:
      #           status:
      #             type: string
      #             example: "cancelled"
      #             description: New status for the booking.
      #           is_deleted:
      #             type: boolean
      #             example: true
      #             description: Flag to indicate soft deletion.
      responses:
        '200':
          description: OK - Booking soft deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 200
                status: "Success"
                message: "Booking has been soft deleted successfully."
                data:
                  id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                  status: "cancelled"
                  deleted_at: "2026-03-20T12:00:00Z"
                request_id: "7f3a2b1c-4d5e-6f70-8a91-b2c3d4e5f678"
                timestamp: "2026-03-20T12:00:05Z"
        '404':
          description: Not Found - Booking ID does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 404
                status: "error"
                message: "Booking not found."
                data: null
                request_id: "a1b2c3d4-e5f6-7890-abcd-1234567890ef"
                timestamp: "2026-03-20T12:05:00Z"
        '422':
          description: Validation Error - Invalid ID or body content.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                code: 422
                status: "error"
                message: "Validation failed"
                error:
                  detail:
                    - loc: ["body", "status"]
                      msg: "Invalid status value"
                      type: "value_error.invalid"
                request_id: "b1a2c3d4-e5f6-7890-abcd-1234567890ef"
                timestamp: "2026-03-20T12:10:00Z"   


components:
  schemas:
    APIResponse:
      title: APIResponse
      type: object
      properties:
        code:
          type: integer
          example: 400
        status:
          type: string
          example: Error 
        message:
          type: string
          example: Validation Error 
        data:
          nullable: true
        error:
          type: object
          nullable: true
          $ref: '#/components/schemas/ValidationError'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
      required:
        - code
        - status
        - message
        - request_id
        - timestamp
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
    ChatRequest:
      type: object
      properties:
        customer_id:
          type: string
          description: Unique identifier for the customer
          example: "cust_001"
        message:
          type: string
          description: The customer's message or question
          example: "Tell me about your top 5 drinks."
      required:
        - customer_id
        - message

    ChatResponse:
      type: object
      properties:
        reply:
          type: string
          description: The chatbot's friendly response
        structured_output:
          type: object
          nullable: true
          description: Structured data for system integration (e.g., billing)
          properties:
            item:
              type: string
            quantity:
              type: integer
            price:
              type: integer
              description: Price in cents
        
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        aws_request_id:
          type: string
          nullable: true
          description: AWS request ID for debugging (if applicable)